requirements.txt
fastapi==0.115.0
pydantic==2.9.2
langchain==0.3.2
langchain-openai==0.2.2
python-dotenv
uvicorn